objId is size_t
git-svn-id: file:///home/roberto/Git/svn@7139 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: d0c620d8cf49b7a43d234ade932e0946548a33a4

+++ trunk/xmipp/libraries/reconstruction/angular_continuous_assign.cpp
@@ -155,7 +155,7 @@ void ProgAngularContinuousAssign::produce_side_info()
 }
 
 // Predict =================================================================
-void ProgAngularContinuousAssign::processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId)
+void ProgAngularContinuousAssign::processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId)
 {
     // Read the image and take its angles from the Metadata
     // if they are available. If not, take them from the header.
